[python] move header polymorphic-convertible to main include dir, expose convertibles for Multibody dynamics
