+++ bindings/python/include/aligator/python/polymorphic-convertible.hpp
